requirements.txt
dotenv==0.9.9
python-dotenv==1.1.0
flask==3.1.0
openai==1.3.0
numpy==2.2.5
PyPDF2==3.0.1
python-docx==1.1.0
pytesseract==0.3.10
Pillow==10.0.0
services==0.1.1
utils==1.0.2
config==0.5.1
pymongo==4.6.0
pydantic==2.11.4
sentence-transformers==4.1.0